fix(bus,dma): escritas estreitas entregam a palavra inteira ao periférico
O barramento dos periféricos não tem byte-enable. Numa escrita de 8 ou 16 bits
para um registrador de 32 bits, o hardware latcha os 32 bits que o CPU está
gravando, e não só a parte endereçada — escrever um byte em `DMA0_ADDR` grava o
endereço completo. Nós descartávamos os bits altos antes de chegar ao bus, e as
escritas estreitas para o bloco de DMA simplesmente sumiam.

A largura é do periférico, não do acesso: SPU, SIO, timers e IRQ são de 16
bits, então lá uma escrita de byte latcha a meia-palavra; o CD-ROM é de 8.

Quatro erros que o `cpu/io-access-bitwidth` apontou junto:

- `JOY_MODE` guardava os 16 bits inteiros; só o fator de baudrate, o tamanho do
  caractere, a paridade e a polaridade do clock existem no silício.
- Uma leitura larga do bloco do CD-ROM devolvia um byte solto em vez do mesmo
  byte replicado nas quatro posições.
- `MDECSTAT` reportava 0xFFFF em repouso, onde o console reporta zero.
- Os flags de `DICR` subiam em toda transferência, mesmo com o canal sem
  interrupção habilitada — deixando bits acesos que o software nunca limpa,
  porque ele só reconhece os canais que pediu para observar.

Seis dos nove registradores divergentes agora batem exatamente com o console.
Faltam `GPUSTAT` e a região de Expansion 3.

+++ crates/psx-core/src/cpu/r3000a.rs
@@ -658,8 +658,8 @@ impl Cpu {
 
         let value = self.reg(instruction.rt());
         match width {
-            LoadWidth::Byte => bus.store8(address, value as u8),
-            LoadWidth::Half => bus.store16(address, value as u16),
+            LoadWidth::Byte => bus.store8(address, value),
+            LoadWidth::Half => bus.store16(address, value),
             LoadWidth::Word => bus.store32(address, value),
         }
     }
